bug fix: reduction in IDOR false positives.
Previously, I mistakenly thought that we could reduce false positives of
expected DORs by checking to see if the result looks the exact same for both
actors. However, I realize that fundamentally, you can't tell the difference
between IDOR and DOR, and will need to rely on whitelisting to support this.