Add exclusions decorator
This adds two decorators
`fuzz_lightyear.exclusions.non_vulnerable_operations`
and `fuzz_lightyear.exclusions.operations`. The former
specifies operations which should not be tested for
vulnerability, and the latter specifies operations
which should not be included in the fuzzing process at all.

This fixture borrows the same code-flow as the factory
fixtures, which is that users use these decorators to
decorate functions which are called by fuzz_lightyear at
a later point.