bug fix: fuzzed headers overwrote auth headers.
Auth headers are provided using a mechanism that isn't tied to the
parameters that the Swagger operation requires. However, in the case
that the Swagger operation requires parameters, the fuzzed input will
override auth headers. This is not desirable.

In this fix, we specifically watch for this case, and address it
accordingly.